requirements.txt
affine==2.2.1
alembic==1.16.1
attrs==25.3.0
blosc2==2.5.1
boto3==1.34.67
botocore==1.34.162
catboost==1.0.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
cmaes==0.11.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.1
decorator==4.4.2
eli5==0.13.0
exceptiongroup==1.3.0
fonttools==4.58.0
future==1.0.0
GDAL==3.6.4
geopandas==1.0.1
graphviz==0.20.3
greenlet==3.2.2
hyperopt==0.2.5
idna==3.10
imageio==2.9.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.4
mlens==0.2.3
mpi4py==4.0.3
msgpack==1.1.0
narwhals==1.40.0
ndindex==1.10.0
networkx==2.5.1
numba==0.60.0
numexpr==2.10.2
numpy==1.25.1
nvidia-nccl-cu12==2.26.5
optuna==3.2.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.1.1
pluggy==1.6.0
py-cpuinfo==9.0.0
py4j==0.10.9.9
pyaml==25.1.0
PyContracts==1.7.9
PyKrige==1.7.0
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.6.1
pyshp==2.1.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.2.0
PyYAML==6.0.2
rasterio==1.3.7
requests==2.32.2
revrand==1.0.0
s3transfer==0.10.4
scikit-image==0.19.1
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.13.1
seaborn==0.13.0
shap==0.45.0
shapely==2.0.7
six==1.17.0
slicer==0.0.7
snuggs==1.4.7
SQLAlchemy==2.0.41
tables==3.9.2
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uncover-ml==0.6.0
urllib3==1.26.20
vecstack==0.4.0
xgboost==2.1.4
zipp==3.21.0
